feat: enhance item card functionality and loading states
- Updated chip component to improve focus visibility for clickable variants.
- Replaced Spinner with ItemCardSkeletonList in multiple pages (ItemArchive, Items, ItemsTrash, ListItems, Member, ShoppingList) for better loading experience.
- Introduced ItemCardMenu component for item actions (edit, archive, restore, delete) with confirmation dialog.
- Added ItemStatus component to manage item status updates with visual feedback and tour hints.
