+++ src/pages/Items.tsx
@@ -4,8 +4,7 @@ import { useSnackbar } from '../lib/snackbar';
 import { useGetItems } from '../lib/useSupabase';
 
 import ItemCreate from '../components/ItemCreate';
-import ItemCard from '../components/ItemCard';
-import { Spinner } from '../components/ui/spinner';
+import ItemCard, { ItemCardSkeletonList } from '../components/ItemCard';
 
 export default function Items() {
 	const { enqueueSnackbar } = useSnackbar();
@@ -36,11 +35,7 @@ export default function Items() {
 					)}
 				</div>
 
-				{isLoading && (
-					<div className='mt-32 flex justify-center'>
-						<Spinner size={32} />
-					</div>
-				)}
+				{isLoading && <ItemCardSkeletonList />}
 			</div>
 
 			<ItemCreate />
